+++ containers/docker-compose-integration-tests.yml
@@ -60,7 +60,7 @@ services:
     entrypoint: >
       /bin/sh -c '\
         sleep 5 && \
-        LOCKBOOK_DEBUG=1 cargo test --release --no-fail-fast -- --nocapture \
+        LOCKBOOK_DEBUG=1 cargo test --release --no-fail-fast --all -- --nocapture \
       '
 
   server_tests:
