+++ core/tests/account_service_tests.rs
@@ -2,7 +2,6 @@ mod integration_test;
 
 #[cfg(test)]
 mod account_tests {
-    use crate::integration_test::{generate_account, random_username, test_config, test_db};
     use lockbook_core::client::ApiError;
     use lockbook_core::repo::account_repo::AccountRepo;
     use lockbook_core::repo::file_metadata_repo::FileMetadataRepo;
@@ -14,6 +13,7 @@ mod account_tests {
     };
     use lockbook_models::account::Account;
     use lockbook_models::api::NewAccountError;
+    use test_utils::{generate_account, random_username, test_config, test_db};
 
     #[test]
     fn create_account_successfully() {
