+++ core/tests/get_usage_tests.rs
@@ -2,7 +2,6 @@ mod integration_test;
 
 #[cfg(test)]
 mod get_usage_tests {
-    use crate::integration_test::{generate_account, random_filename, test_config};
     use lockbook_core::repo::document_repo::DocumentRepo;
     use lockbook_core::{
         create_account, create_file, delete_file, get_root, get_usage, init_logger, sync_all,
@@ -11,6 +10,7 @@ mod get_usage_tests {
     use lockbook_models::file_metadata::FileType;
     use lockbook_models::file_metadata::FileType::Folder;
     use std::path::Path;
+    use test_utils::{generate_account, random_filename, test_config};
 
     #[macro_export]
     macro_rules! sync_all {
