+++ utils/test_utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+authors = ["Parth <parth@mehrotra.me>"]
+edition = "2018"
+
+[dependencies]
+lockbook-core = { path = "../../core" }
+lockbook-models = { path = "../../core/libs/models" }
+lockbook-crypto = { path = "../../core/libs/crypto" }
+uuid = { version = "0.8.1", features = ["v4", "serde"] }
+serde_json = "1.0.44"
+serde = { version = "1.0", features = ["derive"] }
